requirements.txt
coreapi
Django
django-filter
django-rest-swagger
djangorestframework
Werkzeug
